requirements.txt
streamlit
requests
feedparser
python-dotenv
beautifulsoup4
